+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Not-released
+
+- Fix Carto colors.
+
 ## 0.3.0
 
 ### ⚠️ Breaking Changes️️️️️️️️️️️ ⚠️